Move lib/ modules into app/
