+++ app/routes.js
@@ -7,7 +7,69 @@ var join = require('path').join;
 exports.router = new express.Router();
 
 var routes = [].concat.apply([], config.apps.map(add_app_routes));
-exports.url_for = require('../lib/url_for')(routes);
+
+exports.url_for = function (view_name, args) {
+
+  for (var i = 0, l = routes.length; i < l; i++) {
+    var route = routes[i];
+
+    if (route.name === view_name) {
+      try {
+        return build_url(route.pattern, args);
+
+      } catch (error) {
+        throw 'url_for("' + view_name + '") failed: ' + error;
+      }
+    }
+  }
+
+  throw 'route not found for ' + view_name;
+};
+
+
+function build_url(pattern, args) {
+
+  var query = Object.assign({}, args);
+  var placeholders = /(\/:\w+\??)/g;
+
+  function matching_argument(match, placeholder) {
+    var name = placeholder.replace(/[/:?]/g, '');
+    var value = placeholder;
+
+    if (!(name in args)) {
+      throw 'missing argument "' + name + '"';
+    }
+
+    delete query[name];
+
+    return '/' + args[name];
+  }
+
+  var url = pattern.replace(placeholders, matching_argument);
+
+  if (Object.keys(query).length && url.indexOf('?') === -1) {
+    url += '?';
+  }
+
+  return url + url_params(query);
+}
+
+
+function url_params(query) {
+  var pairs = [];
+
+  for (var property in query) {
+    if (query.hasOwnProperty(property)) {
+
+      var key = encodeURIComponent(property);
+      var val = encodeURIComponent(query[property]);
+
+      pairs.push(key + '=' + val);
+    }
+  }
+
+  return pairs.join('&');
+}
 
 
 function add_app_routes(app_name) {
